refactor(sdk): Unify the attest chain and registry maps into one deployments table
One entry per chain carries the viem config and, where deployed, the
registry address. The viem-free attest-registries subpath is gone with
its only consumer: the button no longer fail-fasts on unsupported
chains — the popup reports the same error through the protocol.
