+++ packages/zkpassport-sdk/src/attest/index.ts
@@ -2,7 +2,7 @@ import type { Chain, PublicClient, WalletClient } from "viem"
 import { createPublicClient, encodeAbiParameters, getAbiItem, http } from "viem"
 import { getChainFromId } from "@zkpassport/utils"
 import type { FacematchMode, NullifierType, ProofResult } from "@zkpassport/utils"
-import { getAttestRegistry } from "./registries"
+import { getAttestRegistry } from "./deployments"
 import { SolidityVerifier } from "../solidity-verifier"
 import { ZKPassportCredentialsAbi } from "../assets/abi/zkpassport-credentials"
 import { PolicyEvaluatorV1Abi } from "../assets/abi/policy-evaluator-v1"
